requirements.txt
streamlit
python-dotenv
requests
edge-tts
google-generativeai
